fix: add forceCallback argument to onLayoutMaybeChanged to manually force calling onLayoutChange
for some reason, when allowOverlap is true the equality check between old layout and new layout
always fails inside the onDragStop handler which leads to onLayoutChange not being called.
On JOOR's The Edit's implementation this causes a bug where the layout between different pages
are incorrectly "in sync".

The fix